drivers/nhs-nut.c: fix compiler complaints [#2692]
        nhs-nut.c: In function ‘get_bit_in_position’:
        nhs-nut.c:357:22: error: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Werror=sign-compare]
          357 |     if (bit_position >= size * 8) {
              |                      ^~

+++ drivers/nhs-nut.c
@@ -348,7 +348,7 @@ static pkt_data lastpktdata = {
     0xFE  // end_marker
 };
 
-int get_bit_in_position(void *ptr, size_t size, int bit_position,int invertorder) {
+int get_bit_in_position(void *ptr, size_t size, size_t bit_position, int invertorder) {
     unsigned char *byte_ptr = (unsigned char *)ptr;
     int retval = -2;
     size_t byte_index = bit_position / 8;
